Made some changes to snakes_and_ladders_analysis
I had problems running the file. Changed the all_moves zip to a list so it could be run in DataFrame. Changes some names in the pyplot histogram that didnt make sense to me, and didnt run for me. Added some labels to x axis and y axis in the pyplot histogram.